Task #5830: revert most KVpair changes. The extra overloads clash on either 32 or 64 bit. Adding overloads will also fail, as they cannot be mapped to PVSS data types in the PVSSGateway (default case there is to try string). So overloads -> template won't work either. Agreed with Ruud to also remove the vector<int> support (already not mapped). Then to solve the original data type problem, adapt a few logged data types from uint or ulong to int. While at it, adapt a few to bool from string ("0" or "1").

+++ trunk/RTCP/Cobalt/GPUProc/src/cuda/Pipelines/Pipeline.cc
@@ -93,7 +93,7 @@ namespace LOFAR
       itsMdLogger.log(itsMdKeyPrefix + PN_CGP_OBSERVATION_NAME, boost::lexical_cast<string>(ps.observationID()));
       for (unsigned i = 0; i < subbandIndices.size(); ++i) {
         itsMdLogger.log(itsMdKeyPrefix + PN_CGP_SUBBAND + '[' + boost::lexical_cast<string>(subbandIndices[i]) + ']',
-                        subbandIndices[i]);
+                        (int)subbandIndices[i]);
       }
     }
